fixed some compile-time errors
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@4706 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ libs/csengine/particle.cpp
@@ -348,7 +348,7 @@ void csSpiralParticleSystem::Update (cs_time elapsed_time)
 
     SetSpeed (part_idx, dir);
     SetAccel (part_idx, csVector3 (0));
-    part->MovePosition( -time_before_new_particle / 1000.0 * dir);
+    part->MovePosition( -(float)time_before_new_particle / 1000.0 * dir);
   }
   csNewtonianParticleSystem::Update (elapsed_time);
 }
